+++ packages/engine-multi/src/engine.ts
@@ -98,11 +98,6 @@ const createEngine = async (options: EngineOptions, workerPath?: string) => {
   const contexts: Record<string, ExecutionContext> = {};
   const deferredListeners: Record<string, Record<string, EventHandler>[]> = {};
 
-  // TODO I think this is for later
-  //const activeWorkflows: string[] = [];
-
-  // TOOD I wonder if the engine should a) always accept a worker path
-  // and b) validate it before it runs
   let resolvedWorkerPath;
   if (workerPath) {
     // If a path to the worker has been passed in, just use it verbatim
@@ -183,14 +178,17 @@ const createEngine = async (options: EngineOptions, workerPath?: string) => {
       delete deferredListeners[workflowId];
     }
 
-    // execute(context);
-
     // Run the execute on a timeout so that consumers have a chance
     // to register listeners
     setTimeout(() => {
       // TODO typing between the class and interface isn't right
       // @ts-ignore
-      execute(context);
+      execute(context).finally(() => {
+        delete contexts[workflowId];
+        if (options.purge && Object.keys(contexts).length === 0) {
+          engine.purge?.();
+        }
+      });
     }, 1);
 
     // hmm. Am I happy to pass the internal workflow state OUT of the handler?
@@ -202,7 +200,6 @@ const createEngine = async (options: EngineOptions, workerPath?: string) => {
         context.once(evt, fn),
       off: (evt: string, fn: (...args: any[]) => void) => context.off(evt, fn),
     };
-    // return context;
   };
 
   const listen = (