+++ packages/engine-multi/src/errors.ts
@@ -57,5 +57,17 @@ export class CompileError extends EngineError {
   }
 }
 
-// Autoinstall Error (exception)
+export class AutoinstallError extends EngineError {
+  severity = 'exception'; // Syntax errors are crashes, but what if we get a module resolution thing?
+  type = 'AutoinstallError';
+  name = 'AutoinstallError';
+  message;
+
+  constructor(specifier: string, error: any) {
+    super();
+
+    this.message = `Error installing ${specifier}: ${error.message}`;
+  }
+}
+
 // CredentialsError (exception)